Added ASCII map level and win detection to PlayScreen
The PlayScreen class now loads levels from a ASCII map using the new Level class, then it sets up the player at the start cell, and detects when the player reaches the exit cell which then triggers an overlay to temporarily represent a win. The Camera and viewport inits are now based on the loaded grid size.