Add StepListener to Player
Added the StepListener interface and updated Player class to accept StepListenerl. Refactored PlayScreen to use StepListener for handling player movement events, then simplifying the update logic. This is to fix a bug where the player is able to enter in the cell where the tether trail was before it fully entered the center of the cell and thus since it did not fully update in the new cell that cell was not considered a solid to the player and thus the player was able to go backwards and inside of the tether and once the tether fully updated in the new cell while the player is inside of it, it results in the player becoming fully stuck inside of the cell that the tether was in.