Added door and switch systems with composite blockers
Added a DoorSystem and SwitchSystem to manage doors and switches in the game. Added CompositeBlocker to allow multiple CellBlockers to be combined, and updates pathfinding, enemy, player, and projectile logic to respec blockers such as doors and tethers. PlayScreen was refactored to use these systems. Enabling doors is togglable by switches via projectiles